Move printers to CLI namespace
